backend/requirements.txt
flask
flask-cors
opencv-python
numpy
python-dotenv
ultralytics
redis
deep-sort-realtime
pyzbar
pandas
torch
torchvision
Pillow
git+https://github.com/openai/CLIP.git
